Make domdec a proper module
Now has
* Doxygen group definition
* initial Doxygen comments
* correct #include guards
* no export "C" for header used only in C++ code
* all relevant current files are declared as such in the module
* new domdec_constraints.h and domdec_internal.h
* split off domdec_specatomcomm.cpp and domdec_vsite.c from
  domdec_constraints.cpp
