+++ src/gromacs/domdec/domdec.h
@@ -32,9 +32,31 @@
  * To help us fund GROMACS development, we humbly ask that you cite
  * the research papers on the package. Check out http://www.gromacs.org.
  */
+/*! \defgroup module_domdec Spatial domain decomposition (for parallelization over MPI)
+ * \ingroup group_mdrun
+ *
+ * \brief Manages the decomposition of the simulation volume over MPI
+ * ranks to try to distribute work evenly with minimal communication
+ * overheads.
+ *
+ * \todo Get domdec stuff out of legacyheaders/types/commrec.h
+ *
+ * \author Berk Hess <hess@kth.se>
+ *
+ */
 
-#ifndef _domdec_h
-#define _domdec_h
+/*! \libinternal \file
+ *
+ * \brief This file declares functions for mdrun to call to manage the
+ * details of its domain decomposition.
+ *
+ * \author Berk Hess <hess@kth.se>
+ * \inlibraryapi
+ * \ingroup module_domdec
+ */
+
+#ifndef GMX_DOMDEC_DOMDEC_H
+#define GMX_DOMDEC_DOMDEC_H
 
 #include <stdio.h>
 
@@ -60,57 +82,55 @@
 extern "C" {
 #endif
 
-int ddglatnr(gmx_domdec_t *dd, int i);
-/* Returns the global topology atom number belonging to local atom index i.
- * This function is intended for writing ascii output
+/*! \brief Returns the global topology atom number belonging to local atom index i.
+ *
+ * This function is intended for writing ASCII output
  * and returns atom numbers starting at 1.
  * When dd=NULL returns i+1.
  */
+int ddglatnr(gmx_domdec_t *dd, int i);
 
+/*! \brief Return a block struct for the charge groups of the whole system */
 t_block *dd_charge_groups_global(gmx_domdec_t *dd);
-/* Return a block struct for the charge groups of the whole system */
 
-gmx_bool dd_filled_nsgrid_home(gmx_domdec_t *dd);
-/* Is the ns grid already filled with the home particles? */
-
-void dd_store_state(gmx_domdec_t *dd, t_state *state);
-/* Store the global cg indices of the home cgs in state,
- * so it can be reset, even after a new DD partitioning.
+/*! \brief Store the global cg indices of the home cgs in state,
+ *
+ * This means it can be reset, even after a new DD partitioning.
  */
+void dd_store_state(gmx_domdec_t *dd, t_state *state);
 
+/*! \brief Returns a pointer to the gmx_domdec_zones_t struct */
 gmx_domdec_zones_t *domdec_zones(gmx_domdec_t *dd);
 
+/*! \brief Sets the j-charge-group range for i-charge-group \p icg */
 void dd_get_ns_ranges(gmx_domdec_t *dd, int icg,
                       int *jcg0, int *jcg1, ivec shift0, ivec shift1);
 
+/*! \brief Returns the atom range in the local state for atoms involved in virtual sites */
 int dd_natoms_vsite(gmx_domdec_t *dd);
 
+/*! \brief Sets the atom range for atom in the local state for atoms received in constraints communication */
 void dd_get_constraint_range(gmx_domdec_t *dd,
                              int *at_start, int *at_end);
 
-real dd_cutoff_mbody(gmx_domdec_t *dd);
-
-real dd_cutoff_twobody(gmx_domdec_t *dd);
-
+/*! \brief Get the number of PME nodes along x and y, can be called with dd=NULL */
 void get_pme_nnodes(const gmx_domdec_t *dd,
                     int *npmenodes_x, int *npmenodes_y);
-/* Get the number of PME nodes along x and y, can be called with dd=NULL */
 
-gmx_bool gmx_pmeonlynode(t_commrec *cr, int nodeid);
-/* Return if nodeid in cr->mpi_comm_mysim is a PME-only node */
-
+/*! \brief Returns the set of DD nodes that communicate with pme node cr->nodeid */
 void get_pme_ddnodes(t_commrec *cr, int pmenodeid,
                      int *nmy_ddnodes, int **my_ddnodes, int *node_peer);
-/* Returns the set of DD nodes that communicate with pme node cr->nodeid */
 
+/*! \brief Returns the maximum shift for coordinate communication in PME, dim x */
 int dd_pme_maxshift_x(gmx_domdec_t *dd);
-/* Returns the maximum shift for coordinate communication in PME, dim x */
 
+/*! \brief Returns the maximum shift for coordinate communication in PME, dim y */
 int dd_pme_maxshift_y(gmx_domdec_t *dd);
-/* Returns the maximum shift for coordinate communication in PME, dim y */
 
+/*! \brief Generates the MPI communicators for domain decomposition */
 void make_dd_communicators(FILE *fplog, t_commrec *cr, int dd_node_order);
 
+/*! \brief Initialized the domain decomposition, chooses the DD grid and PME ranks */
 gmx_domdec_t *
 init_domain_decomposition(FILE *fplog,
                           t_commrec *cr,
@@ -124,92 +144,111 @@ init_domain_decomposition(FILE *fplog,
                           gmx_ddbox_t *ddbox,
                           int *npme_x, int *npme_y);
 
+/*! \brief Initialize data structures for bonded interactions */
 void dd_init_bondeds(FILE *fplog,
                      gmx_domdec_t *dd, gmx_mtop_t *mtop,
                      gmx_vsite_t *vsite,
                      t_inputrec *ir, gmx_bool bBCheck, cginfo_mb_t *cginfo_mb);
-/* Initialize data structures for bonded interactions */
 
+/*! \brief Returns if we need to do pbc for calculating bonded interactions */
 gmx_bool dd_bonded_molpbc(gmx_domdec_t *dd, int ePBC);
-/* Returns if we need to do pbc for calculating bonded interactions */
 
+/*! \brief Set DD grid dimensions and limits
+ *
+ * Should be called after calling dd_init_bondeds.
+ */
 void set_dd_parameters(FILE *fplog, gmx_domdec_t *dd, real dlb_scale,
                        t_inputrec *ir,
                        gmx_ddbox_t *ddbox);
-/* Set DD grid dimensions and limits,
- * should be called after calling dd_init_bondeds.
- */
 
-gmx_bool change_dd_cutoff(t_commrec *cr, t_state *state, t_inputrec *ir,
-                          real cutoff_req );
-/* Change the DD non-bonded communication cut-off.
+/*! \brief Change the DD non-bonded communication cut-off.
+ *
  * This could fail when trying to increase the cut-off,
  * then FALSE will be returned and the cut-off is not modified.
  */
+gmx_bool change_dd_cutoff(t_commrec *cr, t_state *state, t_inputrec *ir,
+                          real cutoff_req );
 
-void change_dd_dlb_cutoff_limit(t_commrec *cr);
-/* Domain boundary changes due to the DD dynamic load balancing can limit
+/*! \brief Limit DLB to preserve the option of returning to the current cut-off.
+ *
+ * Domain boundary changes due to the DD dynamic load balancing can limit
  * the cut-off distance that can be set in change_dd_cutoff. This function
  * limits the DLB such that using the currently set cut-off should still be
  * possible after subsequently setting a shorter cut-off with change_dd_cutoff.
  */
+void change_dd_dlb_cutoff_limit(t_commrec *cr);
 
+/*! \brief Return if the DLB lock is set */
 gmx_bool dd_dlb_is_locked(const gmx_domdec_t *dd);
-/* Return if the DLB lock is set */
 
+/*! \brief Set a lock such that with DLB=auto DLB can (not) get turned on */
 void dd_dlb_set_lock(gmx_domdec_t *dd, gmx_bool bValue);
-/* Set a lock such that with DLB=auto DLB can (not) get turned on */
 
-void dd_setup_dlb_resource_sharing(t_commrec           *cr,
-                                   const gmx_hw_info_t *hwinfo,
-                                   const gmx_hw_opt_t  *hw_opt);
-/* When domains (PP MPI ranks) share a GPU, the individual GPU wait times
+/*! \brief Set up communication for averaging GPU wait times over ranks
+ *
+ * When domains (PP MPI ranks) share a GPU, the individual GPU wait times
  * are meaningless, as it depends on the order in which tasks on the same
  * GPU finish. Therefore there wait times need to be averaged over the ranks
  * sharing the same GPU. This function sets up the communication for that.
  */
+void dd_setup_dlb_resource_sharing(t_commrec           *cr,
+                                   const gmx_hw_info_t *hwinfo,
+                                   const gmx_hw_opt_t  *hw_opt);
 
+/*! \brief Sets up the DD communication setup */
 void setup_dd_grid(FILE *fplog, gmx_domdec_t *dd);
 
+/*! \brief Collects local rvec arrays \p lv to \p v on the master rank */
 void dd_collect_vec(gmx_domdec_t *dd,
                     t_state *state_local, rvec *lv, rvec *v);
 
+/*! \brief Collects the local state \p state_local to \p state on the master rank */
 void dd_collect_state(gmx_domdec_t *dd,
                       t_state *state_local, t_state *state);
 
+/*! \brief Cycle counter indices used internally in the domain decomposition */
 enum {
     ddCyclStep, ddCyclPPduringPME, ddCyclF, ddCyclWaitGPU, ddCyclPME, ddCyclNr
 };
 
+/*! \brief Add the wallcycle count to the DD counter */
 void dd_cycles_add(gmx_domdec_t *dd, float cycles, int ddCycl);
-/* Add the wallcycle count to the DD counter */
 
+/*! \brief Start the force flop count */
 void dd_force_flop_start(gmx_domdec_t *dd, t_nrnb *nrnb);
-/* Start the force flop count */
 
+/*! \brief Stop the force flop count */
 void dd_force_flop_stop(gmx_domdec_t *dd, t_nrnb *nrnb);
-/* Stop the force flop count */
 
-float dd_pme_f_ratio(gmx_domdec_t *dd);
-/* Return the PME/PP force load ratio, or -1 if nothing was measured.
+/*! \brief Return the PME/PP force load ratio, or -1 if nothing was measured.
+ *
  * Should only be called on the DD master node.
  */
+float dd_pme_f_ratio(gmx_domdec_t *dd);
 
+/*! \brief Communicate the coordinates to the neighboring cells and do pbc. */
 void dd_move_x(gmx_domdec_t *dd, matrix box, rvec x[]);
-/* Communicate the coordinates to the neighboring cells and do pbc. */
 
-void dd_move_f(gmx_domdec_t *dd, rvec f[], rvec *fshift);
-/* Sum the forces over the neighboring cells.
+/*! \brief Sum the forces over the neighboring cells.
+ *
  * When fshift!=NULL the shift forces are updated to obtain
  * the correct virial from the single sum including f.
  */
+void dd_move_f(gmx_domdec_t *dd, rvec f[], rvec *fshift);
 
+/*! \brief Communicate a real for each atom to the neighboring cells. */
 void dd_atom_spread_real(gmx_domdec_t *dd, real v[]);
-/* Communicate a real for each atom to the neighboring cells. */
 
+/*! \brief Sum the contributions to a real for each atom over the neighboring cells. */
 void dd_atom_sum_real(gmx_domdec_t *dd, real v[]);
-/* Sum the contributions to a real for each atom over the neighboring cells. */
 
+/*! \brief Partition the system over the nodes.
+ *
+ * step is only used for printing error messages.
+ * If bMasterState==TRUE then state_global from the master node is used,
+ * else state_local is redistributed between the nodes.
+ * When f!=NULL, *f will be reallocated to the size of state_local.
+ */
 void dd_partition_system(FILE                *fplog,
                          gmx_int64_t          step,
                          t_commrec           *cr,
@@ -229,62 +268,52 @@ void dd_partition_system(FILE                *fplog,
                          t_nrnb              *nrnb,
                          gmx_wallcycle_t      wcycle,
                          gmx_bool             bVerbose);
-/* Partition the system over the nodes.
- * step is only used for printing error messages.
- * If bMasterState==TRUE then state_global from the master node is used,
- * else state_local is redistributed between the nodes.
- * When f!=NULL, *f will be reallocated to the size of state_local.
- */
 
+/*! \brief Reset all the statistics and counters for total run counting */
 void reset_dd_statistics_counters(gmx_domdec_t *dd);
-/* Reset all the statistics and counters for total run counting */
 
+/*! \brief Print statistics for domain decomposition communication */
 void print_dd_statistics(t_commrec *cr, t_inputrec *ir, FILE *fplog);
 
 /* In domdec_con.c */
 
+/*! \brief Communicates the virtual site forces, reduces the shift forces when \p fshift != NULL */
 void dd_move_f_vsites(gmx_domdec_t *dd, rvec *f, rvec *fshift);
 
+/*! \brief Clears the forces for virtual sites */
 void dd_clear_f_vsites(gmx_domdec_t *dd, rvec *f);
 
+/*! \brief Move x0 and also x1 if x1!=NULL. bX1IsCoord tells if to do PBC on x1 */
 void dd_move_x_constraints(gmx_domdec_t *dd, matrix box,
                            rvec *x0, rvec *x1, gmx_bool bX1IsCoord);
-/* Move x0 and also x1 if x1!=NULL. bX1IsCoord tells if to do PBC on x1 */
 
+/*! \brief Communicates the coordinates involved in virtual sites */
 void dd_move_x_vsites(gmx_domdec_t *dd, matrix box, rvec *x);
 
+/*! \brief Returns the local atom count array for all constraints
+ *
+ * The local atom count for a constraint, possible values 2/1/0, is needed
+ * to avoid not/double-counting contributions linked to the Lagrange
+ * multiplier, such as the virial and free-energy derivatives.
+ */
 int *dd_constraints_nlocalatoms(gmx_domdec_t *dd);
-
-void dd_clear_local_constraint_indices(gmx_domdec_t *dd);
-
-void dd_clear_local_vsite_indices(gmx_domdec_t *dd);
-
-int dd_make_local_vsites(gmx_domdec_t *dd, int at_start, t_ilist *lil);
-
-int dd_make_local_constraints(gmx_domdec_t *dd, int at_start,
-                              const gmx_mtop_t *mtop,
-                              const int *cginfo,
-                              gmx_constr_t constr, int nrec,
-                              t_ilist *il_local);
-
-void init_domdec_constraints(gmx_domdec_t *dd,
-                             gmx_mtop_t   *mtop);
-
-void init_domdec_vsites(gmx_domdec_t *dd, int n_intercg_vsite);
-
 
 /* In domdec_top.c */
 
+/*! \brief Print error output when interactions are missing */
 void dd_print_missing_interactions(FILE *fplog, t_commrec *cr,
                                    int local_count,  gmx_mtop_t *top_global, t_state *state_local);
 
+/*! \brief Generate and store the reverse topology */
 void dd_make_reverse_top(FILE *fplog,
                          gmx_domdec_t *dd, gmx_mtop_t *mtop,
                          gmx_vsite_t *vsite,
                          t_inputrec *ir, gmx_bool bBCheck);
 
+/*! \brief Store the local charge group index in \p lcgs */
 void dd_make_local_cgs(gmx_domdec_t *dd, t_block *lcgs);
 
+/*! \brief Generate the local topology and virtual site data */
 void dd_make_local_top(gmx_domdec_t *dd, gmx_domdec_zones_t *zones,
                        int npbcdim, matrix box,
                        rvec cellsize_min, ivec npulse,
@@ -293,56 +322,64 @@ void dd_make_local_top(gmx_domdec_t *dd, gmx_domdec_zones_t *zones,
                        gmx_vsite_t *vsite,
                        gmx_mtop_t *top, gmx_localtop_t *ltop);
 
+/*! \brief Sort ltop->ilist when we are doing free energy. */
 void dd_sort_local_top(gmx_domdec_t *dd, t_mdatoms *mdatoms,
                        gmx_localtop_t *ltop);
-/* Sort ltop->ilist when we are doing free energy. */
 
+/*! \brief Construct local topology */
 gmx_localtop_t *dd_init_local_top(gmx_mtop_t *top_global);
 
+/*! \brief Construct local state */
 void dd_init_local_state(gmx_domdec_t *dd,
                          t_state *state_global, t_state *local_state);
 
+/*! \brief Generate a list of links between charge groups that are linked by bonded interactions */
 t_blocka *make_charge_group_links(gmx_mtop_t *mtop, gmx_domdec_t *dd,
                                   cginfo_mb_t *cginfo_mb);
 
+/*! \brief Calculate the maximum distance involved in 2-body and multi-body bonded interactions */
 void dd_bonded_cg_distance(FILE *fplog, gmx_mtop_t *mtop,
                            t_inputrec *ir, rvec *x, matrix box,
                            gmx_bool bBCheck,
                            real *r_2b, real *r_mb);
 
+/*! \brief Dump a pdb file with the current DD home + communicated atoms.
+ *
+ * When natoms=-1, dump all known atoms.
+ */
 void write_dd_pdb(const char *fn, gmx_int64_t step, const char *title,
                   gmx_mtop_t *mtop,
                   t_commrec *cr,
                   int natoms, rvec x[], matrix box);
-/* Dump a pdb file with the current DD home + communicated atoms.
- * When natoms=-1, dump all known atoms.
- */
 
 
 /* In domdec_setup.c */
 
+/*! \brief Returns the volume fraction of the system that is communicated */
 real comm_box_frac(ivec dd_nc, real cutoff, gmx_ddbox_t *ddbox);
-/* Returns the volume fraction of the system that is communicated */
 
+/*! \brief Determines the optimal DD cell setup dd->nc and possibly npmenodes
+ * for the system.
+ *
+ * On the master node returns the actual cellsize limit used.
+ */
 real dd_choose_grid(FILE *fplog,
                     t_commrec *cr, gmx_domdec_t *dd, t_inputrec *ir,
                     gmx_mtop_t *mtop, matrix box, gmx_ddbox_t *ddbox,
                     gmx_bool bDynLoadBal, real dlb_scale,
                     real cellsize_limit, real cutoff_dd,
                     gmx_bool bInterCGBondeds);
-/* Determines the optimal DD cell setup dd->nc and possibly npmenodes
- * for the system.
- * On the master node returns the actual cellsize limit used.
- */
 
 
 /* In domdec_box.c */
 
+/*! \brief Set the box and PBC data in \p ddbox */
 void set_ddbox(gmx_domdec_t *dd, gmx_bool bMasterState, t_commrec *cr_sum,
                t_inputrec *ir, matrix box,
                gmx_bool bCalcUnboundedSize, t_block *cgs, rvec *x,
                gmx_ddbox_t *ddbox);
 
+/*! \brief Set the box and PBC data in \p ddbox */
 void set_ddbox_cr(t_commrec *cr, ivec *dd_nc,
                   t_inputrec *ir, matrix box, t_block *cgs, rvec *x,
                   gmx_ddbox_t *ddbox);
@@ -351,4 +388,4 @@ void set_ddbox_cr(t_commrec *cr, ivec *dd_nc,
 }
 #endif
 
-#endif  /* _domdec_h */
+#endif
